Update project structure - generate clear packages for architecture - extend gitignore - add test package
