+++ src/main/model/theWorld/DungeonBuilder.java
@@ -1,4 +1,4 @@
-package model_AdventureGame.theWorld;
+package model.theWorld;
 
 import java.util.ArrayList;
 import java.util.Random;
